move breakpoints stuff together into shared
