+++ apps/builder/app/builder/features/breakpoints/index.ts
@@ -1,3 +1,4 @@
 export * from "./breakpoints-settings";
 export * from "./breakpoints-selector";
 export * from "./use-set-initial-canvas-width";
+export * from "./group-breakpoints";
