+++ apps/builder/app/builder/features/breakpoints/trigger-button.tsx
@@ -8,7 +8,7 @@ import {
   toggleItemStyle,
 } from "@webstudio-is/design-system";
 import { scaleStore, useCanvasWidth } from "~/builder/shared/nano-states";
-import { selectedBreakpointStore } from "~/shared/nano-states/breakpoints";
+import { selectedBreakpointStore } from "~/builder/shared/breakpoints/stores";
 
 const Value = ({
   children,
