+++ apps/builder/app/builder/features/breakpoints/use-set-initial-canvas-width.ts
@@ -5,7 +5,7 @@ import {
   workspaceRectStore,
 } from "~/builder/shared/nano-states";
 import { breakpointsStore } from "~/shared/nano-states";
-import { selectedBreakpointStore } from "~/shared/nano-states/breakpoints";
+import { selectedBreakpointStore } from "~/builder/shared/breakpoints/stores";
 import { findInitialWidth } from "./find-initial-width";
 
 export const useSetInitialCanvasWidth = () => {
