+++ apps/builder/app/builder/features/breakpoints/width-setting.tsx
@@ -6,7 +6,7 @@ import { breakpointsStore } from "~/shared/nano-states";
 import {
   selectedBreakpointIdStore,
   selectedBreakpointStore,
-} from "~/shared/nano-states/breakpoints";
+} from "~/builder/shared/breakpoints/stores";
 
 // Doesn't make sense to allow resizing the canvas lower/higher than this.
 export const minWidth = 240;
