+++ apps/builder/app/builder/shared/breakpoints/stores.ts
@@ -1,7 +1,6 @@
 import { atom, computed } from "nanostores";
 import type { Breakpoint } from "@webstudio-is/project-build";
-import { breakpointsStore } from "./nano-states";
-import { compareMedia } from "@webstudio-is/css-engine";
+import { breakpointsStore } from "../../../shared/nano-states/nano-states";
 import { isBaseBreakpoint } from "~/builder/shared/breakpoints";
 
 export const selectedBreakpointIdStore = atom<undefined | Breakpoint["id"]>(
@@ -25,20 +24,6 @@ export const selectedBreakpointStore = computed(
     );
   }
 );
-
-/**
- * order number starts with 1 and covers all existing breakpoints
- */
-export const selectBreakpointByOrderNumber = (orderNumber: number) => {
-  const breakpoints = breakpointsStore.get();
-  const index = orderNumber - 1;
-  const breakpoint = Array.from(breakpoints.values())
-    .sort(compareMedia)
-    .at(index);
-  if (breakpoint) {
-    selectedBreakpointIdStore.set(breakpoint.id);
-  }
-};
 
 export const synchronizedBreakpointsStores = [
   ["selectedBreakpointId", selectedBreakpointIdStore],
